# go-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A comprehensive anonymous form builder and response collection system",
        "title": "AnoQ Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "AnoQ Support",
            "email": "support@anoq.com"
        },
        "license": {
            "name": "MIT",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/form/{id}/questions": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Add a new question to an existing form",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Questions"
                ],
                "summary": "Create a question for a form",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Form ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Question data",
                        "name": "question",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateQuestionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Question created successfully",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "question": {
                                    "$ref": "#/definitions/model.QuestionResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request body or form ID",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied: you don't own this form",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Form not found",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/questions/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get details of a specific question",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Questions"
                ],
                "summary": "Get a question by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Question ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Question details",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "question": {
                                    "$ref": "#/definitions/model.QuestionResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid question ID",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied: you don't own this form",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Question not found",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.CreateQuestionRequest": {
            "description": "Request payload for creating a new question",
            "type": "object",
            "required": [
                "question_text",
                "type"
            ],
            "properties": {
                "allow_multiple": {
                    "description": "Allow multiple selections for multiple_choice questions",
                    "type": "boolean",
                    "example": false
                },
                "choices": {
                    "description": "Choices for multiple_choice questions",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"Very satisfied\"",
                        " \"Satisfied\"",
                        " \"Neutral\"]"
                    ]
                },
                "position": {
                    "description": "Position in form (optional, auto-assigned if not provided)",
                    "type": "integer",
                    "example": 1
                },
                "question_text": {
                    "description": "Question text (required)",
                    "type": "string",
                    "example": "How satisfied are you with our service?"
                },
                "required": {
                    "description": "Whether question is required",
                    "type": "boolean",
                    "example": true
                },
                "type": {
                    "description": "Question type: basic or multiple_choice (required)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.QuestionType"
                        }
                    ],
                    "example": "multiple_choice"
                }
            }
        },
...
        "model.QuestionResponse": {
            "type": "object",
            "properties": {
                "allow_multiple": {
                    "type": "boolean"
                },
                "answer": {
                    "type": "string"
                },
                "choices": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "form_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "position": {
                    "type": "integer"
                },
                "question_text": {
                    "type": "string"
                },
                "required": {
                    "type": "boolean"
                },
                "selected_choice": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "type": {
                    "$ref": "#/definitions/model.QuestionType"
                }
            }
        },
        "model.QuestionType": {
            "type": "string",
            "enum": [
                "basic",
                "multiple_choice"
            ],
            "x-enum-varnames": [
                "QuestionTypeBasic",
                "QuestionTypeMultipleChoice"
            ]
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "SessionToken": {
            "description": "Session token for authentication",
            "type": "apiKey",
            "name": "X-Session-Token",
            "in": "header"
        }
    }
}